requirements.txt
pyqt4
pymediainfo
psutil
hidapi
youtube-dl
numpy